Support INDI 1.9.0.
Support INDI::PropertyView as a replacement of concrete structs. I
think that we need to move to using the accessor methods insted of
directly touching the underlying fields. This would be a bigger
change, so it might call for a bump to version 0.3.0.

I've also added an udpated vega.py, from the tutorial, as a good way
to better exercise the API.